Reduce number of calls to Localstack
